+++ iati/default.py
@@ -16,27 +16,29 @@ import iati.constants
 import iati.resources
 
 
-def get_default_version_if_none(version):
-    """Return the default version number if the input version is None. Otherwise returns the input version as is.
+def _specific_version_for(version):
+    """Convert a version number into the most appropriate Decimal Version.
+
+    * Decimal Versions will remain unchanged.
+    * Integer Versions will return the latest Decimal Version within the Integer.
 
     Args:
-        version (str or None): The version to test against.
+        version (str): The version to obtain a specific version for.
 
     Raises:
-        ValueError: When a specified version is not a valid version of the IATI Standard.
+        ValueError: When a specified version cannot be converted to a valid version of the IATI Standard.
 
     Returns:
-        str: The default version if the input version is None. Otherwise returns the input version.
-
-    Note:
-        The default version of the Standard is deemed to be the latest version.
+        str: The Decimal Version of the Standard that the input version relates to.
 
     """
-    if version is None:
-        return iati.constants.STANDARD_VERSION_LATEST
-    elif version not in iati.constants.STANDARD_VERSIONS:
-        raise ValueError("Version {0} is not a valid version of the IATI Standard.".format(version))
-    return version
+    if version in iati.constants.STANDARD_VERSIONS:
+        return version
+    elif version in [str(v) for v in iati.constants.STANDARD_VERSIONS_MAJOR]:
+        decimals_for_integer = [v for v in iati.constants.STANDARD_VERSIONS if v.split('.')[0] == version]
+        return max(decimals_for_integer)
+    else:
+        raise ValueError('Version {0} is not a valid version of the IATI Standard.'.format(version))
 
 
 _CODELISTS = defaultdict(dict)
@@ -64,12 +66,12 @@ Warning:
 """
 
 
-def codelist(name, version=None):
+def codelist(name, version):
     """Return the default Codelist with the specified name for the specified version of the Standard.
 
     Args:
         name (str): The name of the Codelist to return.
-        version (str): The version of the Standard to return the Codelists for. Defaults to None. This means that a version-independent version of the specified Codelist is returned.
+        version (str): The Integer or Decimal version of the Standard to return the Codelist for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
 
     Raises:
         ValueError: When a specified name is not a Codelist at the specified version of the Standard.
@@ -98,11 +100,11 @@ def codelist(name, version=None):
         raise ValueError(msg)
 
 
-def _codelists(version=None, use_cache=False):
+def _codelists(version, use_cache=False):
     """Locate the default Codelists for the specified version of the Standard.
 
     Args:
-        version (str): The version of the Standard to return the Codelists for. Defaults to None. This means that a version-independent version of the Codelists are returned.
+        version (str): The Integer or Decimal version of the Standard to return the Codelists for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
         use_cache (bool): Whether the cache should be used rather than loading the Codelists from disk again. If used, a `deepcopy()` should be performed on any returned Codelist before it is modified.
 
     Raises:
@@ -119,7 +121,7 @@ def _codelists(version=None, use_cache=False):
         This is a private function so as to prevent the (dangerous) `use_cache` parameter being part of the public API.
 
     """
-    version = get_default_version_if_none(version)
+    version = _specific_version_for(version)
 
     paths = iati.resources.get_all_codelist_paths(version)
 
@@ -134,11 +136,11 @@ def _codelists(version=None, use_cache=False):
     return _CODELISTS[version]
 
 
-def codelists(version=None):
+def codelists(version):
     """Return the default Codelists for the specified version of the Standard.
 
     Args:
-        version (str): The version of the Standard to return the Codelists for. Defaults to None. This means that a version-independent version of the Codelists are returned.
+        version (str): The Integer or Decimal version of the Standard to return the Codelists for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
 
     Raises:
         ValueError: When a specified version is not a valid version of the IATI Standard.
@@ -150,11 +152,11 @@ def codelists(version=None):
     return _codelists(version)
 
 
-def codelist_mapping(version=None):
+def codelist_mapping(version):
     """Define the mapping process which states where in a Dataset you should find values on a given Codelist.
 
     Args:
-        version (str): The version of the Standard to return the mapping file for. Defaults to None. This means that a version-independent mapping file is returned.
+        version (str): The Integer or Decimal version of the Standard to return the Codelist Mapping File for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
 
     Raises:
         ValueError: When a specified version is not a valid version of the IATI Standard.
@@ -166,6 +168,8 @@ def codelist_mapping(version=None):
         Make use of the `version` parameter.
 
     """
+    version = _specific_version_for(version)
+
     path = iati.resources.get_codelist_mapping_path(version)
     mapping_tree = iati.utilities.load_as_tree(path)
     mappings = defaultdict(list)
@@ -187,11 +191,11 @@ def codelist_mapping(version=None):
     return mappings
 
 
-def ruleset(version=None):
+def ruleset(version):
     """Return the Standard Ruleset for the specified version of the Standard.
 
     Args:
-        version (str): The version of the Standard to return the Ruleset for. Defaults to None. This means that a version-independent Standard Ruleset is returned.
+        version (str): The Integer or Decimal version of the Standard to return the Standard Ruleset for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
 
     Raises:
         ValueError: When a specified version is not a valid version of the IATI Standard.
@@ -200,17 +204,16 @@ def ruleset(version=None):
         iati.Ruleset: The default Ruleset for the specified version of the Standard.
 
     """
+    version = _specific_version_for(version)
+
     path = iati.resources.get_ruleset_path(iati.resources.FILE_RULESET_STANDARD_NAME, version)
     ruleset_str = iati.utilities.load_as_string(path)
 
     return iati.Ruleset(ruleset_str)
 
 
-def ruleset_schema(version=None):
+def ruleset_schema():
     """Return the Ruleset schema for the specified version of the Standard.
-
-    Args:
-        version (str): The version of the Standard to return the Ruleset for. Defaults to None. This means that a version-independent Ruleset schema is returned.
 
     Raises:
         ValueError: When a specified version is not a valid version of the IATI Standard.
@@ -218,8 +221,11 @@ def ruleset_schema(version=None):
     Returns:
         dict: A dictionary representing the Ruleset schema for the specified version of the Standard.
 
+    Todo:
+        Determine whether a version should be provided. This is worth considering if the content of the IATI Ruleset Schema varies between versions.
+
     """
-    path = iati.resources.get_ruleset_path(iati.resources.FILE_RULESET_SCHEMA_NAME, version)
+    path = iati.resources.get_ruleset_path(iati.resources.FILE_RULESET_SCHEMA_NAME)
     schema_str = iati.utilities.load_as_string(path)
 
     return json.loads(schema_str)
@@ -253,14 +259,14 @@ Warning:
 """
 
 
-def _populate_schema(schema, version=None):
+def _populate_schema(schema, version):
     """Populate a Schema with all its extras.
 
     The extras include Codelists and Rulesets.
 
     Args:
         schema (iati.Schema): The Schema to populate.
-        version (str): The version of the Standard to return the Schema for. Defaults to None. This means that a version-independent version of the Standard is assumed.
+        version (str): The Integer or Decimal version of the Standard to return the Schema for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
 
     Returns:
         iati.Schema: The provided Schema, populated with additional information.
@@ -269,7 +275,7 @@ def _populate_schema(schema, version=None):
         Does not create a copy of the provided Schema, instead adding to it directly.
 
     """
-    version = get_default_version_if_none(version)
+    version = _specific_version_for(version)
 
     codelists_to_add = codelists(version)
     for codelist_to_add in codelists_to_add.values():
@@ -280,13 +286,13 @@ def _populate_schema(schema, version=None):
     return schema
 
 
-def _schema(path_func, schema_class, version=None, populate=True, use_cache=False):
+def _schema(path_func, schema_class, version, populate=True, use_cache=False):
     """Return the default Schema of the specified type for the specified version of the Standard.
 
     Args:
         path_func (func): A function to return the paths at which the relevant Schema can be found.
         schema_class (type): A class definition for the Schema of interest.
-        version (str): The version of the Standard to return the Schema for. Defaults to None. This means that a version-independent version of the Schema is returned.
+        version (str): The Integer or Decimal version of the Standard to return the Schema for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
         populate (bool): Whether the Schema should be populated with auxilliary information such as Codelists and Rulesets.
         use_cache (bool): Whether the cache should be used rather than loading the Schema from disk again. If used, a `deepcopy()` should be performed on any returned Schema before it is modified.
 
@@ -299,7 +305,7 @@ def _schema(path_func, schema_class, version=None, populate=True, use_cache=Fals
     """
     population_key = 'populated' if populate else 'unpopulated'
 
-    version = get_default_version_if_none(version)
+    version = _specific_version_for(version)
 
     schema_paths = path_func(version)
 
@@ -312,11 +318,11 @@ def _schema(path_func, schema_class, version=None, populate=True, use_cache=Fals
     return _SCHEMAS[version][population_key][schema_class.ROOT_ELEMENT_NAME]
 
 
-def activity_schema(version=None, populate=True):
+def activity_schema(version, populate=True):
     """Return the default Activity Schema for the specified version of the Standard.
 
     Args:
-        version (str): The version of the Standard to return the Schema for. Defaults to None. This means that a version-independent version of the Schema is returned.
+        version (str): The Integer or Decimal version of the Standard to return the Activity Schema for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
         populate (bool): Whether the Schema should be populated with auxilliary information such as Codelists and Rulesets.
 
     Raises:
@@ -329,11 +335,11 @@ def activity_schema(version=None, populate=True):
     return _schema(iati.resources.get_all_activity_schema_paths, iati.ActivitySchema, version, populate)
 
 
-def organisation_schema(version=None, populate=True):
+def organisation_schema(version, populate=True):
     """Return the default Organisation Schema for the specified version of the Standard.
 
     Args:
-        version (str): The version of the Standard to return the Schema for. Defaults to None. This means that a version-independent version of the Schema is returned.
+        version (str): The Integer or Decimal version of the Standard to return the Organisation Schema for. If an Integer Version is specified, uses the most recent Decimal Version within the Integer Version.
         populate (bool): Whether the Schema should be populated with auxilliary information such as Codelists and Rulesets.
 
     Raises: